bsd.cabal.mk: Work around a probable bug in pkg(8)
pkg(8) keeps the last file in the packaging list open for writing when
executing the post-install actions.  Thus, when this last file is a binary
which is to be executed on post-install, it will fail with ETXTBSY.  And, as a
result, the post-install script will also fail.  For example, this happens
with devel/hs-haddock, where the gen_contents_index script wants to invoke the
freshly installed version of haddock at post-install.

Solve this problem by moving the binaries in the package list up.