Add a (default) xmonad workaround for losing hotkeys in certain cases.
This has also been reported in the xmonad issue tracker [1], where Dominik
posted the patch has now been added to the port.  So far, our joint
investigations led to the conclusion that this has something to do with the
differences in handling interval timers on FreeBSD.  Unfortunately, we could
not yet identify the exact source of the bug therefore we cannot provide a
fix for the run-time system yet -- but until it is fixed, adding a workaround
for the xmonad port itself looks a reasonable compromise.

This fix becomes a default option because it does no harm to the systems where
this issue is not observed.

PR:			181049
Submitted by:	Dominik Ernst <de@dernst.org>

[1] https://code.google.com/p/xmonad/issues/detail?id=576