lang/ghc: Remove the superfluous dependency on misc/compat9x.
This also unbreaks the build in the install phase when the port is built with
synth.  Actually, misc/compat9x is not needed at all by the 9.x-targeted
bootstrap compiler and as marino noted in the PR this line did not have any
effect anyway but caused troubles.

Submitted by:	Walter Schwarzenfeld <w.litter@aon.at>
PR:			206794